+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Rgba8, Gray8 pixel formats.
 * Raster now has a pixel format type parameter.
 * Raster::width(), height() and pixels()
+* Raster::with_pixels() (for borrowed pixels).
 ### Removed
 * Plotter::over, raster and write_png methods.
 * Plotter no longer has an associated Raster